Allow multiply where parameters

+++ tests.php
@@ -187,6 +187,13 @@ if ($db->count != 2) {
     exit;
 }
 
+$db->where("id = ? or id = ?", Array(1,2));
+$res = $db->get ("users");
+if ($db->count != 2) {
+    echo "Invalid users count on select with multiple params";
+    exit;
+}
+
 $db->delete("users");
 $db->get("users");
 if ($db->count != 0) {
